Reorganise spatial aggregates to live under location
