+++ flowmachine/flowmachine/core/server/query_schemas/spatial_aggregate.py
@@ -6,11 +6,10 @@ from marshmallow import Schema, fields, post_load
 from marshmallow.validate import OneOf
 from marshmallow_oneofschema import OneOfSchema
 
-
-from flowmachine.features.utilities.spatial_aggregates import (
-    SpatialAggregate,
+from flowmachine.features.location.redacted_spatial_aggregate import (
     RedactedSpatialAggregate,
 )
+from flowmachine.features.location.spatial_aggregate import SpatialAggregate
 from .base_exposed_query import BaseExposedQuery
 from .daily_location import DailyLocationSchema
 from .modal_location import ModalLocationSchema
