+++ flowmachine/flowmachine/features/location/joined_spatial_aggregate.py
@@ -1,96 +1,9 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-"""
-Utility classes for performing spatial aggregate
-operations in CDRs.
-"""
 import warnings
 from typing import List
 
-from ...core.query import Query
-from ...core.mixins import GeoDataMixin
+from flowmachine.core import Query
+from flowmachine.core.mixins import GeoDataMixin
 from flowmachine.utils import parse_datestring
-
-
-class SpatialAggregate(GeoDataMixin, Query):
-    """
-    Class representing the result of spatially aggregating
-    a locations object. A locations object represents the
-    location of multiple subscribers. This class represents the output
-    of aggregating that data spatially.
-
-    Parameters
-    ----------
-    locations : subscriber location query
-    """
-
-    def __init__(self, *, locations):
-
-        self.locations = locations
-        # self.spatial_unit is used in self._geo_augmented_query
-        self.spatial_unit = locations.spatial_unit
-        super().__init__()
-
-    @property
-    def column_names(self) -> List[str]:
-        return self.spatial_unit.location_id_columns + ["total"]
-
-    def _make_query(self):
-
-        aggregate_cols = ",".join(self.spatial_unit.location_id_columns)
-
-        sql = f"""
-        SELECT
-            {aggregate_cols},
-            count(*) AS total
-        FROM
-            ({self.locations.get_query()}) AS to_agg
-        GROUP BY
-            {aggregate_cols}
-        """
-
-        return sql
-
-
-class RedactedSpatialAggregate(GeoDataMixin, Query):
-    """
-    Class representing the result of spatially aggregating
-    a locations object, redacted so that results are not returned if counts are 15 or less..
-    A locations object represents the
-    location of multiple subscribers. This class represents the output
-    of aggregating that data spatially.
-
-    Parameters
-    ----------
-    locations : subscriber location query
-    """
-
-    def __init__(self, *, spatial_aggregate: SpatialAggregate):
-
-        self.spatial_aggregate = spatial_aggregate
-        # self.spatial_unit is used in self._geo_augmented_query
-        self.spatial_unit = spatial_aggregate.spatial_unit
-        super().__init__()
-
-    @property
-    def column_names(self) -> List[str]:
-        return self.spatial_unit.location_id_columns + ["total"]
-
-    def _make_query(self):
-
-        aggregate_cols = ",".join(self.spatial_aggregate.columns)
-
-        sql = f"""
-        SELECT
-            {aggregate_cols}
-        FROM
-            ({self.spatial_aggregate.get_query()}) AS agged
-        WHERE agged.total > 15
-        """
-
-        return sql
 
 
 class JoinedSpatialAggregate(GeoDataMixin, Query):
