+++ flowmachine/flowmachine/features/location/redacted_spatial_aggregate.py
@@ -1,0 +1,44 @@
+from typing import List
+
+from flowmachine.core import Query
+from flowmachine.core.mixins import GeoDataMixin
+from flowmachine.features.location.spatial_aggregate import SpatialAggregate
+
+
+class RedactedSpatialAggregate(GeoDataMixin, Query):
+    """
+    Class representing the result of spatially aggregating
+    a locations object, redacted so that results are not returned if counts are 15 or less..
+    A locations object represents the
+    location of multiple subscribers. This class represents the output
+    of aggregating that data spatially.
+
+    Parameters
+    ----------
+    locations : subscriber location query
+    """
+
+    def __init__(self, *, spatial_aggregate: SpatialAggregate):
+
+        self.spatial_aggregate = spatial_aggregate
+        # self.spatial_unit is used in self._geo_augmented_query
+        self.spatial_unit = spatial_aggregate.spatial_unit
+        super().__init__()
+
+    @property
+    def column_names(self) -> List[str]:
+        return self.spatial_unit.location_id_columns + ["total"]
+
+    def _make_query(self):
+
+        aggregate_cols = ",".join(self.spatial_aggregate.columns)
+
+        sql = f"""
+        SELECT
+            {aggregate_cols}
+        FROM
+            ({self.spatial_aggregate.get_query()}) AS agged
+        WHERE agged.total > 15
+        """
+
+        return sql
